Changes on review feedback.
