+++ logging/api/LoggingSample.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QuickStart", "QuickStart\Qu
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LoggingTest", "LoggingTest\LoggingTest.csproj", "{DEBA8E76-4986-4C6C-957E-C77FCD910474}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WithGitId", "WithGitId\WithGitId.csproj", "{8683659B-6463-46BF-B427-24AFDD6BB227}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{DEBA8E76-4986-4C6C-957E-C77FCD910474}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DEBA8E76-4986-4C6C-957E-C77FCD910474}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DEBA8E76-4986-4C6C-957E-C77FCD910474}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8683659B-6463-46BF-B427-24AFDD6BB227}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8683659B-6463-46BF-B427-24AFDD6BB227}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8683659B-6463-46BF-B427-24AFDD6BB227}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8683659B-6463-46BF-B427-24AFDD6BB227}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
