+++ logging/api/WithGitId/Properties/AssemblyInfo.cs
