+++ logging/api/WithGitId/Views/LoggingSample/Index.cshtml
