+++ logging/api/WithGitId/Web.Release.config
