+++ logging/api/WithGitId/Web.config
