+++ logging/api/WithGitId/packages.config
